moved handleLike
